9C-1141-3 Refactor Applications-Like Processes
We refactor the `processes` module. We add a new `applications`
package, for those processes that retrieve or use information about
Android Applications.

There were three Processes classes, Application~, Recommendations~,
and GooglePlay~, that were too closely related. We have merged them
and put then into the `applications` package. We also merge their
tests.
