+++ modules/api/src/main/scala/cards/nine/app/AppResolverActor.scala
@@ -6,7 +6,7 @@ import cards.nine.commons.NineCardsService._
 import cards.nine.commons.TaskInstances._
 import cards.nine.commons.config.NineCardsConfig._
 import cards.nine.domain.application.ResolvePendingStats
-import cards.nine.processes.ApplicationProcesses
+import cards.nine.processes.applications.ApplicationProcesses
 import cats.~>
 import scalaz.{ -\/, \/, \/- }
 import scalaz.concurrent.Task
