+++ modules/processes/src/main/scala/cards/nine/processes/applications/Processes.scala
@@ -1,12 +1,43 @@
-package cards.nine.processes
+package cards.nine.processes.applications
 
 import cards.nine.commons.NineCardsService
+import cards.nine.commons.NineCardsService.NineCardsService
 import cards.nine.commons.NineCardsService._
 import cards.nine.domain.application.{ BasicCard, CardList, FullCard, Package, PriceFilter }
+import cards.nine.domain.application.{ ResolvePendingStats }
 import cards.nine.domain.market.MarketCredentials
 import cards.nine.services.free.algebra.GooglePlay
 
-class RecommendationsProcesses[F[_]](implicit services: GooglePlay.Services[F]) {
+class ApplicationProcesses[F[_]](implicit services: GooglePlay.Services[F]) {
+
+  import Converters._
+
+  def getAppsInfo(
+    packagesName: List[Package],
+    marketAuth: MarketCredentials
+  ): NineCardsService[F, CardList[FullCard]] =
+    if (packagesName.isEmpty)
+      NineCardsService.right(CardList(Nil, Nil))
+    else
+      services.resolveManyDetailed(
+        packageNames = packagesName,
+        auth         = marketAuth
+      ) map filterCategorized
+
+  def getAppsBasicInfo(
+    packagesName: List[Package],
+    marketAuth: MarketCredentials
+  ): NineCardsService[F, CardList[BasicCard]] =
+    if (packagesName.isEmpty)
+      NineCardsService.right(CardList(Nil, Nil))
+    else
+      services.resolveManyBasic(packagesName, marketAuth)
+
+  def resolvePendingApps(numPackages: Int): NineCardsService[F, ResolvePendingStats] =
+    services.resolvePendingApps(numPackages)
+
+  def storeCard(card: FullCard): NineCardsService[F, Unit] =
+    services.storeCard(card)
 
   def getRecommendationsByCategory(
     category: String,
@@ -55,9 +86,9 @@ class RecommendationsProcesses[F[_]](implicit services: GooglePlay.Services[F]) 
     )
 }
 
-object RecommendationsProcesses {
+object ApplicationProcesses {
 
-  implicit def recommendationsProcesses[F[_]](implicit services: GooglePlay.Services[F]) =
-    new RecommendationsProcesses
+  implicit def applicationProcesses[F[_]](implicit services: GooglePlay.Services[F]) =
+    new ApplicationProcesses
 
 }
